compilation: support custom baudrates in musl-based compilers
For some reason, the code in ser-unix checks for the TCGETS2 without a
proper include to the kernel header <asm/ioctls.h>. Due to this, the
code uses the regular termios struct in a non POSIX compliant way, which
musl does not accept.

To resolve this, i simply added an include for <asm/ioctls.h>.

NOTE: some architecture do not contain the TCGETS2 ioctl in their kernel
headers. The solution would be to use the hidden fields in musl's
declaration, but they incompatible with glibc's (they have __ prefix).

Since i would like to preserve gcc compilation capability in this repo,
and since musl detection is not a thing, i simply removed support for custom
baudrates for non asm-generic architecures (in my gdb-static repo,
powerpc is an example of such architecture). I assume/hope that no one
needs this brand-new feature of gdb on such old architectures.